GSIN-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic[email]==2.8.2
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.43
python-multipart==0.0.9
python-dotenv==1.0.1
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
email-validator==2.1.0
bcrypt==4.1.2
python-jose==3.3.0
cryptography==42.0.5
stripe==10.0.0
httpx==0.27.0
psutil==5.9.8
sentry-sdk==2.15.0
sentence-transformers==2.7.0
pytest==8.3.3
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0
redis==5.0.1
alpaca-trade-api>=3.1.1
